requirements.txt
flask
requests
openai
torch
torchvision
pillow
